Various cleanup and refactoring.

+++ src/git.rs
@@ -5,25 +5,23 @@ use std::path::Path;
 use errors::*;
 
 pub fn shallow_clone_or_pull(url: &str, dir: &Path) -> Result<()> {
-    let ref url = frob_url(url);
+    let url = frob_url(url);
 
     if !dir.exists() {
         log!("cloning {} into {}", url, dir.display());
         let r = run::run("git",
-                         &["clone", "--depth", "1", url, &dir.to_string_lossy()],
+                         &["clone", "--depth", "1", &url, &dir.to_string_lossy()],
                          &[])
             .chain_err(|| format!("unable to clone {}", url));
 
-        if r.is_err() {
-            if dir.exists() {
-                fs::remove_dir_all(dir)?;
-            }
+        if r.is_err() && dir.exists() {
+            fs::remove_dir_all(dir)?;
         }
 
         r
     } else {
         log!("pulling existing url {} into {}", url, dir.display());
-        run::cd_run(&dir,
+        run::cd_run(dir,
                     "git",
                     &["pull"],
                     &[])
@@ -35,7 +33,7 @@ pub fn shallow_clone_or_pull(url: &str, dir: &Path) -> Result<()> {
 /// first check whether it does, and if not do increasingly deep clones until it
 /// finds the commit.
 pub fn shallow_fetch_sha(url: &str, dir: &Path, sha: &str) -> Result<()> {
-    let ref url = frob_url(url);
+    let url = frob_url(url);
 
     log!("ensuring sha {} in {}", sha, url);
     let depths = &[1, 10, 100, 1000];
@@ -46,11 +44,7 @@ pub fn shallow_fetch_sha(url: &str, dir: &Path, sha: &str) -> Result<()> {
                                 "git",
                                 &["log", sha],
                                 &[]);
-            if r.is_ok() {
-                true
-            } else {
-                false
-            }
+            r.is_ok()
         } else {
             false
         }
@@ -62,7 +56,7 @@ pub fn shallow_fetch_sha(url: &str, dir: &Path, sha: &str) -> Result<()> {
         util::try_hard(|| {
             run::run("git",
                      &["clone", "--depth", &format!("{}", depth),
-                       url, &dir.to_string_lossy()],
+                       &url, &dir.to_string_lossy()],
                      &[])
         }).chain_err(|| format!("unable to clone {}", url))?;
 
@@ -71,7 +65,7 @@ pub fn shallow_fetch_sha(url: &str, dir: &Path, sha: &str) -> Result<()> {
 
     util::try_hard(|| {
             run::run("git",
-                     &["clone", url, &dir.to_string_lossy()],
+                     &["clone", &url, &dir.to_string_lossy()],
                      &[])
     }).chain_err(|| format!("unable to clone {}", url))?;
